go/parser, go/ast: accept multi-dim. index expressions a[i, j, k...]
To represent more than one index in the AST, the ast.IndexExpr node
has to change. It now takes a list of expressions rather than just
one expression in its index field.

This has some repercussions for code that depends on the AST. Where
it is trivial to handle the list we do so (e.g. go/printer), In all
other cases we only look at the first element and complain if there
is more than one index (e.g. go/types). While touching more files
than desired, this change is mostly trivial plumbing and all the
places which needed adjusting were easily found by running all.bash.

Since this is a change to an exported go/ast type, this is not a Go 1
compatible change. Consequently, the API check at the end of all.bash
fails.